'curl': add support for download tool
works perfectly, but with one __very big__ caveat:

* 'curl' is limited in the way the output filepath can be specified
* there are two options:
  1. "-o" "/absolute/path/to/filename.ext"
     which requires specifying both:
       * the directory path
       * the filename
  2. "-O"
     which chooses both of these values:
       * the `current working directory`
         is chosen for the directory path
       * the filename is determined by the URL.
         this is actually a bad methodology,
         but there's a supplemental option:
           "--remote-header-name"
         which takes the value from the "content-disposition" header.

* the problem is:
  * we want to specify the directory path, _but_
  * we want to take the filename to be determined for us

In this commit, the argument passed to 'curl' is:
  "-O" "--remote-header-name"

which causes all downloads to be saved into
the `current working directory`,
which is the directory that contains the Firefox executable.

this is a horrible workflow, _and_
__this__ is why 'curl' hasn't found its way into `master`.

though there's nothing that can be done about selecting the CWD,
I described in the release notes for `v2.00` a strategy
that I'm planning to implement, which will allow a user
to interactively select the complete filepath,
which would allow the 'curl' implementation to change
the command line argument, and instead use the "-o <path>" switch.