[v2.03] new hidden preference to control whether URLs are case sensitive
by default, the following URL-related values will be normalized to lowercase:
  * contextual variables:
      `request.original_uri`
      `request.uri`
      `request.referrer`
  * the request URL that is tested against rule `url` regex patterns

the purpose for this update is to remove the need
to add a `case insensitive` switch to all regex patterns
within the rule data sets.

it should be a (trivial) performance improvement.

note: names of HTTP headers are always normalized to lower case