[v2.04] adds new contextual variable: `request.window_location`
I'm in the process of starting a "recipe book".. so to speak.
Writing and organizing some sample rules, which can be extremely useful.
This collection of data files is going to be assigned its own branch.
I'd like to receive pull requests.. as others share their good ideas.

In any case, I was testing a rule in the response data set..
that would conditionally add permissive Cross-Origin Resource Sharing (CORS) headers,
when the requesting page is either `file:` protocol or `localhost` host.

Using `request.referrer` was getting me nowhere fast.
When the page is loaded from `file:`, the referrer is empty/undefined.
When it loads an external resource (such as a css file),
which in turn loads yet another external resource..
the `referrer` is set to walk these resources,
rather than pointing to the URL of the original page.

Maybe it's supposed to work that way,
but I'm not finding it particularly useful.

Hence, the addition of the new variable: `request.window_location`
which is always extracted from the `location` object of the "most recent" window.

Seems to work very nicely for the particular test case.